Move revision model/manager to its own submodule
